+++ src/main/java/com/t3f4/zerowaste/global/config/SwaggerConfig.java
@@ -1,4 +1,4 @@
-package com.t3f4.zerowaste.config;
+package com.t3f4.zerowaste.global.config;
 
 import io.swagger.v3.oas.models.OpenAPI;
 import io.swagger.v3.oas.models.info.Info;
